model_runner/vits-ft/requirements_v2.txt
absl-py
aiofiles
aiohttp
aiosignal
altair
annotated-types
antlr4-python3-runtime
anyio
asttokens
async-timeout
attrs
audioread
backcall
backports.functools-lru-cache
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
cn2an
colorama
comm
contourpy
cycler
dataclasses-json
debugpy
decorator
demucs
dora_search
einops
eng-to-ipa
et-xmlfile
executing
faiss-cpu
fastapi
ffmpeg-python
ffmpy
filelock
fonttools
frozenlist
fsspec
future
google-auth
google-auth-oauthlib
gradio_client
greenlet
grpcio
h11
httpcore
httpx
huggingface-hub
idna
imageio-ffmpeg
importlib-resources
indic-transliteration
inflect
ipykernel
ipython
jamo
jedi
jieba
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
julius
jupyter_client
jupyter_core
kiwisolver
ko-pron
lameenc
langchain
langchain-experimental
langsmith
lazy_loader
librosa
llvmlite
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
more-itertools
mpmath
msgpack
multidict
mypy-extensions
nest-asyncio
networkx
num-thai
numexpr
oauthlib
omegaconf
OpenCC
openpyxl
openunmix
orjson
packaging
pandas
parso
pickleshare
pillow
platformdirs
pooch
proces
proglog
prompt-toolkit
protobuf
psutil
pure-eval
pyasn1
pyasn1-modules
pycparser
pydantic_core
pydub
Pygments
pyopenjtalk-prebuilt
pyparsing
pypinyin
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
referencing
requests
requests-oauthlib
resampy
retrying
rich
roman
rpds-py
rsa
ruamel.yaml.clib
ruff
scikit-learn
scipy
seaborn
semantic-version
shellingham
six
sniffio
soxr
SQLAlchemy
stack-data
starlette
submitit
sympy
tenacity
tensorboard
tensorboard-data-server
termcolor
threadpoolctl
tiktoken
toml
tomlkit
toolz
torch
torchaudio
torchvision
tornado
tqdm
traitlets
treetable
typer
typing-inspect
typing_extensions
tzdata
Unidecode
urllib3
uvicorn
wcwidth
websockets
Werkzeug
yarl
eng_to_ipa
ko_pron
indic_transliteration
num_thai